pyproject.toml
[project]
name = "sncf-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "slack-sdk>=3.39.0",
    "tzdata>=2025.3",
]
